splitting RNG operations
