+++ ext/random/random.c
@@ -1,0 +1,753 @@
+/*
+   +----------------------------------------------------------------------+
+   | Copyright (c) The PHP Group                                          |
+   +----------------------------------------------------------------------+
+   | This source file is subject to version 3.01 of the PHP license,      |
+   | that is bundled with this package in the file LICENSE, and is        |
+   | available through the world-wide-web at the following url:           |
+   | https://www.php.net/license/3_01.txt                                 |
+   | If you did not receive a copy of the PHP license and are unable to   |
+   | obtain it through the world-wide-web, please send a note to          |
+   | license@php.net so we can mail you a copy immediately.               |
+   +----------------------------------------------------------------------+
+   | Authors: Rasmus Lerdorf <rasmus@php.net>                             |
+   |          Zeev Suraski <zeev@php.net>                                 |
+   |          Sascha Schumann <sascha@schumann.cx>                        |
+   |          Pedro Melo <melo@ip.pt>                                     |
+   |          Sterling Hughes <sterling@php.net>                          |
+   |          Sammy Kaye Powers <me@sammyk.me>                            |
+   |                                                                      |
+   | Based on code from: Richard J. Wagner <rjwagner@writeme.com>         |
+   |                     Makoto Matsumoto <matumoto@math.keio.ac.jp>      |
+   |                     Takuji Nishimura                                 |
+   |                     Shawn Cokus <Cokus@math.washington.edu>          |
+   +----------------------------------------------------------------------+
+*/
+
+/*
+	The following php_mt_...() functions are based on a C++ class MTRand by
+	Richard J. Wagner. For more information see the web page at
+	http://www.math.sci.hiroshima-u.ac.jp/~m-mat/MT/VERSIONS/C-LANG/MersenneTwister.h
+
+	Mersenne Twister random number generator -- a C++ class MTRand
+	Based on code by Makoto Matsumoto, Takuji Nishimura, and Shawn Cokus
+	Richard J. Wagner  v1.0  15 May 2003  rjwagner@writeme.com
+
+	The Mersenne Twister is an algorithm for generating random numbers.  It
+	was designed with consideration of the flaws in various other generators.
+	The period, 2^19937-1, and the order of equidistribution, 623 dimensions,
+	are far greater.  The generator is also fast; it avoids multiplication and
+	division, and it benefits from caches and pipelines.  For more information
+	see the inventors' web page at http://www.math.sci.hiroshima-u.ac.jp/~m-mat/MT/emt.html
+
+	Reference
+	M. Matsumoto and T. Nishimura, "Mersenne Twister: A 623-Dimensionally
+	Equidistributed Uniform Pseudo-Random Number Generator", ACM Transactions on
+	Modeling and Computer Simulation, Vol. 8, No. 1, January 1998, pp 3-30.
+
+	Copyright (C) 1997 - 2002, Makoto Matsumoto and Takuji Nishimura,
+	Copyright (C) 2000 - 2003, Richard J. Wagner
+	All rights reserved.
+
+	Redistribution and use in source and binary forms, with or without
+	modification, are permitted provided that the following conditions
+	are met:
+
+	1. Redistributions of source code must retain the above copyright
+	   notice, this list of conditions and the following disclaimer.
+
+	2. Redistributions in binary form must reproduce the above copyright
+	   notice, this list of conditions and the following disclaimer in the
+	   documentation and/or other materials provided with the distribution.
+
+	3. The names of its contributors may not be used to endorse or promote
+	   products derived from this software without specific prior written
+	   permission.
+
+	THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+	"AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+	LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+	A PARTICULAR PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL THE COPYRIGHT OWNER OR
+	CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+	EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+	PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+	PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+	LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+	NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+	SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+*/
+
+#ifdef HAVE_CONFIG_H
+# include "config.h"
+#endif
+
+#include <stdlib.h>
+#include <sys/stat.h>
+#include <fcntl.h>
+#include <math.h>
+
+#include "php.h"
+#include "Zend/zend_exceptions.h"
+#include "php_random.h"
+
+#if HAVE_UNISTD_H
+# include <unistd.h>
+#endif
+
+#ifdef PHP_WIN32
+# include "win32/time.h"
+# include "win32/winutil.h"
+# include <process.h>
+#else
+# include <sys/time.h>
+#endif
+
+#ifdef __linux__
+# include <sys/syscall.h>
+#endif
+
+#if HAVE_SYS_PARAM_H
+# include <sys/param.h>
+# if (__FreeBSD__ && __FreeBSD_version > 1200000) || (__DragonFly__ && __DragonFly_version >= 500700) || defined(__sun)
+#  include <sys/random.h>
+# endif
+#endif
+
+#if __has_feature(memory_sanitizer)
+# include <sanitizer/msan_interface.h>
+#endif
+
+#include "random_arginfo.h"
+
+/*
+ * combinedLCG() returns a pseudo random number in the range of (0, 1).
+ * The function combines two CGs with periods of
+ * 2^31 - 85 and 2^31 - 249. The period of this function
+ * is equal to the product of both primes.
+ */
+
+#define MODMULT(a, b, c, m, s) q = s/a;s=b*(s-a*q)-c*q;if(s<0)s+=m
+
+#define N             MT_N                 /* length of state vector */
+#define M             (397)                /* a period parameter */
+#define hiBit(u)      ((u) & 0x80000000U)  /* mask all but highest   bit of u */
+#define loBit(u)      ((u) & 0x00000001U)  /* mask all but lowest    bit of u */
+#define loBits(u)     ((u) & 0x7FFFFFFFU)  /* mask     the highest   bit of u */
+#define mixBits(u, v) (hiBit(u)|loBits(v)) /* move hi bit of u to hi bit of v */
+
+#define twist(m,u,v)  (m ^ (mixBits(u,v)>>1) ^ ((uint32_t)(-(int32_t)(loBit(v))) & 0x9908b0dfU))
+#define twist_php(m,u,v)  (m ^ (mixBits(u,v)>>1) ^ ((uint32_t)(-(int32_t)(loBit(u))) & 0x9908b0dfU))
+
+ZEND_DECLARE_MODULE_GLOBALS(random)
+
+/* {{{ lcg_seed */
+static void lcg_seed(void)
+{
+	struct timeval tv;
+
+	if (gettimeofday(&tv, NULL) == 0) {
+		RANDOM_G(lcg_s1) = tv.tv_sec ^ (tv.tv_usec<<11);
+	} else {
+		RANDOM_G(lcg_s1) = 1;
+	}
+#ifdef ZTS
+	RANDOM_G(lcg_s2) = (zend_long) tsrm_thread_id();
+#else
+	RANDOM_G(lcg_s2) = (zend_long) getpid();
+#endif
+
+	/* Add entropy to s2 by calling gettimeofday() again */
+	if (gettimeofday(&tv, NULL) == 0) {
+		RANDOM_G(lcg_s2) ^= (tv.tv_usec<<11);
+	}
+
+	RANDOM_G(lcg_seeded) = 1;
+}
+/* }}} */
+
+/* {{{ php_mt_initialize */
+static inline void php_mt_initialize(uint32_t seed, uint32_t *state)
+{
+	/* Initialize generator state with seed
+	   See Knuth TAOCP Vol 2, 3rd Ed, p.106 for multiplier.
+	   In previous versions, most significant bits (MSBs) of the seed affect
+	   only MSBs of the state array.  Modified 9 Jan 2002 by Makoto Matsumoto. */
+
+	uint32_t *s = state;
+	uint32_t *r = state;
+	int i = 1;
+
+	*s++ = seed & 0xffffffffU;
+	for( ; i < N; ++i ) {
+		*s++ = ( 1812433253U * ( *r ^ (*r >> 30) ) + i ) & 0xffffffffU;
+		r++;
+	}
+}
+/* }}} */
+
+/* {{{ php_mt_reload */
+static inline void php_mt_reload(void)
+{
+	/* Generate N new values in state
+	   Made clearer and faster by Matthew Bellew (matthew.bellew@home.com) */
+
+	uint32_t *state = RANDOM_G(mt_rand_state);
+	uint32_t *p = state;
+	int i;
+
+	if (RANDOM_G(mt_rand_mode) == MT_RAND_MT19937) {
+		for (i = N - M; i--; ++p)
+			*p = twist(p[M], p[0], p[1]);
+		for (i = M; --i; ++p)
+			*p = twist(p[M-N], p[0], p[1]);
+		*p = twist(p[M-N], p[0], state[0]);
+	}
+	else {
+		for (i = N - M; i--; ++p)
+			*p = twist_php(p[M], p[0], p[1]);
+		for (i = M; --i; ++p)
+			*p = twist_php(p[M-N], p[0], p[1]);
+		*p = twist_php(p[M-N], p[0], state[0]);
+	}
+	RANDOM_G(mt_rand_left) = N;
+	RANDOM_G(mt_rand_next) = state;
+}
+/* }}} */
+
+static uint32_t rand_range32(uint32_t umax) {
+	uint32_t result, limit;
+
+	result = php_mt_rand();
+
+	/* Special case where no modulus is required */
+	if (UNEXPECTED(umax == UINT32_MAX)) {
+		return result;
+	}
+
+	/* Increment the max so the range is inclusive of max */
+	umax++;
+
+	/* Powers of two are not biased */
+	if ((umax & (umax - 1)) == 0) {
+		return result & (umax - 1);
+	}
+
+	/* Ceiling under which UINT32_MAX % max == 0 */
+	limit = UINT32_MAX - (UINT32_MAX % umax) - 1;
+
+	/* Discard numbers over the limit to avoid modulo bias */
+	while (UNEXPECTED(result > limit)) {
+		result = php_mt_rand();
+	}
+
+	return result % umax;
+}
+
+#if ZEND_ULONG_MAX > UINT32_MAX
+static uint64_t rand_range64(uint64_t umax) {
+	uint64_t result, limit;
+
+	result = php_mt_rand();
+	result = (result << 32) | php_mt_rand();
+
+	/* Special case where no modulus is required */
+	if (UNEXPECTED(umax == UINT64_MAX)) {
+		return result;
+	}
+
+	/* Increment the max so the range is inclusive of max */
+	umax++;
+
+	/* Powers of two are not biased */
+	if ((umax & (umax - 1)) == 0) {
+		return result & (umax - 1);
+	}
+
+	/* Ceiling under which UINT64_MAX % max == 0 */
+	limit = UINT64_MAX - (UINT64_MAX % umax) - 1;
+
+	/* Discard numbers over the limit to avoid modulo bias */
+	while (UNEXPECTED(result > limit)) {
+		result = php_mt_rand();
+		result = (result << 32) | php_mt_rand();
+	}
+
+	return result % umax;
+}
+#endif
+
+/* {{{ php_combined_lcg */
+PHPAPI double php_combined_lcg(void)
+{
+	int32_t q, z;
+
+	if (!RANDOM_G(lcg_seeded)) {
+		lcg_seed();
+	}
+
+	MODMULT(53668, 40014, 12211, 2147483563L, RANDOM_G(lcg_s1));
+	MODMULT(52774, 40692, 3791, 2147483399L, RANDOM_G(lcg_s2));
+
+	z = RANDOM_G(lcg_s1) - RANDOM_G(lcg_s2);
+	if (z < 1) {
+		z += 2147483562;
+	}
+
+	return z * 4.656613e-10;
+}
+/* }}} */
+
+/* {{{ php_mt_srand */
+PHPAPI void php_mt_srand(uint32_t seed)
+{
+	/* Seed the generator with a simple uint32 */
+	php_mt_initialize(seed, RANDOM_G(mt_rand_state));
+	php_mt_reload();
+
+	/* Seed only once */
+	RANDOM_G(mt_rand_is_seeded) = 1;
+}
+/* }}} */
+
+/* {{{ php_mt_rand */
+PHPAPI uint32_t php_mt_rand(void)
+{
+	/* Pull a 32-bit integer from the generator state
+	   Every other access function simply transforms the numbers extracted here */
+
+	uint32_t s1;
+
+	if (UNEXPECTED(!RANDOM_G(mt_rand_is_seeded))) {
+		zend_long bytes;
+		if (php_random_bytes_silent(&bytes, sizeof(zend_long)) == FAILURE) {
+			bytes = GENERATE_SEED();
+		}
+		php_mt_srand(bytes);
+	}
+
+	if (RANDOM_G(mt_rand_left) == 0) {
+		php_mt_reload();
+	}
+	--RANDOM_G(mt_rand_left);
+
+	s1 = *RANDOM_G(mt_rand_next)++;
+	s1 ^= (s1 >> 11);
+	s1 ^= (s1 <<  7) & 0x9d2c5680U;
+	s1 ^= (s1 << 15) & 0xefc60000U;
+	return ( s1 ^ (s1 >> 18) );
+}
+/* }}} */
+
+/* {{{ php_mt_rand_range */
+PHPAPI zend_long php_mt_rand_range(zend_long min, zend_long max)
+{
+	zend_ulong umax = max - min;
+
+#if ZEND_ULONG_MAX > UINT32_MAX
+	if (umax > UINT32_MAX) {
+		return (zend_long) (rand_range64(umax) + min);
+	}
+#endif
+
+	return (zend_long) (rand_range32(umax) + min);
+}
+/* }}} */
+
+/* {{{ php_mt_rand_common
+ * rand() allows min > max, mt_rand does not */
+PHPAPI zend_long php_mt_rand_common(zend_long min, zend_long max)
+{
+	int64_t n;
+
+	if (RANDOM_G(mt_rand_mode) == MT_RAND_MT19937) {
+		return php_mt_rand_range(min, max);
+	}
+
+	/* Legacy mode deliberately not inside php_mt_rand_range()
+	 * to prevent other functions being affected */
+	n = (int64_t)php_mt_rand() >> 1;
+	RAND_RANGE_BADSCALING(n, min, max, PHP_MT_RAND_MAX);
+
+	return n;
+}
+/* }}} */
+
+/* {{{ php_srand */
+PHPAPI void php_srand(zend_long seed)
+{
+	php_mt_srand(seed);
+}
+/* }}} */
+
+/* {{{ php_rand */
+PHPAPI zend_long php_rand(void)
+{
+	return php_mt_rand();
+}
+/* }}} */
+
+/* {{{ php_random_bytes */
+PHPAPI int php_random_bytes(void *bytes, size_t size, bool should_throw)
+{
+#ifdef PHP_WIN32
+	/* Defer to CryptGenRandom on Windows */
+	if (php_win32_get_random_bytes(bytes, size) == FAILURE) {
+		if (should_throw) {
+			zend_throw_exception(zend_ce_exception, "Could not gather sufficient random data", 0);
+		}
+		return FAILURE;
+	}
+#elif HAVE_DECL_ARC4RANDOM_BUF && ((defined(__OpenBSD__) && OpenBSD >= 201405) || (defined(__NetBSD__) && __NetBSD_Version__ >= 700000001) || defined(__APPLE__))
+	arc4random_buf(bytes, size);
+#else
+	size_t read_bytes = 0;
+	ssize_t n;
+#if (defined(__linux__) && defined(SYS_getrandom)) || (defined(__FreeBSD__) && __FreeBSD_version >= 1200000) || (defined(__DragonFly__) && __DragonFly_version >= 500700) || defined(__sun)
+	/* Linux getrandom(2) syscall or FreeBSD/DragonFlyBSD getrandom(2) function*/
+	/* Keep reading until we get enough entropy */
+	while (read_bytes < size) {
+		/* Below, (bytes + read_bytes)  is pointer arithmetic.
+
+		   bytes   read_bytes  size
+		     |      |           |
+		    [#######=============] (we're going to write over the = region)
+		             \\\\\\\\\\\\\
+		              amount_to_read
+
+		*/
+		size_t amount_to_read = size - read_bytes;
+#if defined(__linux__)
+		n = syscall(SYS_getrandom, bytes + read_bytes, amount_to_read, 0);
+#else
+		n = getrandom(bytes + read_bytes, amount_to_read, 0);
+#endif
+
+		if (n == -1) {
+			if (errno == ENOSYS) {
+				/* This can happen if PHP was compiled against a newer kernel where getrandom()
+				 * is available, but then runs on an older kernel without getrandom(). If this
+				 * happens we simply fall back to reading from /dev/urandom. */
+				ZEND_ASSERT(read_bytes == 0);
+				break;
+			} else if (errno == EINTR || errno == EAGAIN) {
+				/* Try again */
+				continue;
+			} else {
+			    /* If the syscall fails, fall back to reading from /dev/urandom */
+				break;
+			}
+		}
+
+#if __has_feature(memory_sanitizer)
+		/* MSan does not instrument manual syscall invocations. */
+		__msan_unpoison(bytes + read_bytes, n);
+#endif
+		read_bytes += (size_t) n;
+	}
+#endif
+	if (read_bytes < size) {
+		int    fd = RANDOM_G(random_fd);
+		struct stat st;
+
+		if (fd < 0) {
+#if HAVE_DEV_URANDOM
+			fd = open("/dev/urandom", O_RDONLY);
+#endif
+			if (fd < 0) {
+				if (should_throw) {
+					zend_throw_exception(zend_ce_exception, "Cannot open source device", 0);
+				}
+				return FAILURE;
+			}
+			/* Does the file exist and is it a character device? */
+			if (fstat(fd, &st) != 0 ||
+# ifdef S_ISNAM
+					!(S_ISNAM(st.st_mode) || S_ISCHR(st.st_mode))
+# else
+					!S_ISCHR(st.st_mode)
+# endif
+			) {
+				close(fd);
+				if (should_throw) {
+					zend_throw_exception(zend_ce_exception, "Error reading from source device", 0);
+				}
+				return FAILURE;
+			}
+			RANDOM_G(random_fd) = fd;
+		}
+
+		for (read_bytes = 0; read_bytes < size; read_bytes += (size_t) n) {
+			n = read(fd, bytes + read_bytes, size - read_bytes);
+			if (n <= 0) {
+				break;
+			}
+		}
+
+		if (read_bytes < size) {
+			if (should_throw) {
+				zend_throw_exception(zend_ce_exception, "Could not gather sufficient random data", 0);
+			}
+			return FAILURE;
+		}
+	}
+#endif
+
+	return SUCCESS;
+}
+/* }}} */
+
+/* {{{ php_random_int */
+PHPAPI int php_random_int(zend_long min, zend_long max, zend_long *result, bool should_throw)
+{
+	zend_ulong umax;
+	zend_ulong trial;
+
+	if (min == max) {
+		*result = min;
+		return SUCCESS;
+	}
+
+	umax = (zend_ulong) max - (zend_ulong) min;
+
+	if (php_random_bytes(&trial, sizeof(trial), should_throw) == FAILURE) {
+		return FAILURE;
+	}
+
+	/* Special case where no modulus is required */
+	if (umax == ZEND_ULONG_MAX) {
+		*result = (zend_long)trial;
+		return SUCCESS;
+	}
+
+	/* Increment the max so the range is inclusive of max */
+	umax++;
+
+	/* Powers of two are not biased */
+	if ((umax & (umax - 1)) != 0) {
+		/* Ceiling under which ZEND_LONG_MAX % max == 0 */
+		zend_ulong limit = ZEND_ULONG_MAX - (ZEND_ULONG_MAX % umax) - 1;
+
+		/* Discard numbers over the limit to avoid modulo bias */
+		while (trial > limit) {
+			if (php_random_bytes(&trial, sizeof(trial), should_throw) == FAILURE) {
+				return FAILURE;
+			}
+		}
+	}
+
+	*result = (zend_long)((trial % umax) + min);
+	return SUCCESS;
+}
+/* }}} */
+
+/* {{{ Returns a value from the combined linear congruential generator */
+PHP_FUNCTION(lcg_value)
+{
+	ZEND_PARSE_PARAMETERS_NONE();
+
+	RETURN_DOUBLE(php_combined_lcg());
+}
+/* }}} */
+
+/* {{{ Seeds Mersenne Twister random number generator */
+PHP_FUNCTION(mt_srand)
+{
+	zend_long seed = 0;
+	zend_long mode = MT_RAND_MT19937;
+
+	ZEND_PARSE_PARAMETERS_START(0, 2)
+		Z_PARAM_OPTIONAL
+		Z_PARAM_LONG(seed)
+		Z_PARAM_LONG(mode)
+	ZEND_PARSE_PARAMETERS_END();
+
+	if (ZEND_NUM_ARGS() == 0) {
+		if (php_random_bytes_silent(&seed, sizeof(zend_long)) == FAILURE) {
+			seed = GENERATE_SEED();
+		}
+	}
+
+	switch (mode) {
+		case MT_RAND_PHP:
+			RANDOM_G(mt_rand_mode) = MT_RAND_PHP;
+			break;
+		default:
+			RANDOM_G(mt_rand_mode) = MT_RAND_MT19937;
+	}
+
+	php_mt_srand(seed);
+}
+/* }}} */
+
+/* {{{ Returns a random number from Mersenne Twister */
+PHP_FUNCTION(mt_rand)
+{
+	zend_long min, max;
+	int argc = ZEND_NUM_ARGS();
+
+	if (argc == 0) {
+		// genrand_int31 in mt19937ar.c performs a right shift
+		RETURN_LONG(php_mt_rand() >> 1);
+	}
+
+	ZEND_PARSE_PARAMETERS_START(2, 2)
+		Z_PARAM_LONG(min)
+		Z_PARAM_LONG(max)
+	ZEND_PARSE_PARAMETERS_END();
+
+	if (UNEXPECTED(max < min)) {
+		zend_argument_value_error(2, "must be greater than or equal to argument #1 ($min)");
+		RETURN_THROWS();
+	}
+
+	RETURN_LONG(php_mt_rand_common(min, max));
+}
+/* }}} */
+
+/* {{{ Returns the maximum value a random number from Mersenne Twister can have */
+PHP_FUNCTION(mt_getrandmax)
+{
+	ZEND_PARSE_PARAMETERS_NONE();
+
+	/*
+	 * Melo: it could be 2^^32 but we only use 2^^31 to maintain
+	 * compatibility with the previous php_rand
+	 */
+	RETURN_LONG(PHP_MT_RAND_MAX); /* 2^^31 */
+}
+/* }}} */
+
+/* {{{ Returns a random number from Mersenne Twister */
+PHP_FUNCTION(rand)
+{
+	zend_long min, max;
+	int argc = ZEND_NUM_ARGS();
+
+	if (argc == 0) {
+		RETURN_LONG(php_mt_rand() >> 1);
+	}
+
+	ZEND_PARSE_PARAMETERS_START(2, 2)
+		Z_PARAM_LONG(min)
+		Z_PARAM_LONG(max)
+	ZEND_PARSE_PARAMETERS_END();
+
+	if (max < min) {
+		RETURN_LONG(php_mt_rand_common(max, min));
+	}
+
+	RETURN_LONG(php_mt_rand_common(min, max));
+}
+/* }}} */
+
+/* {{{ Return an arbitrary length of pseudo-random bytes as binary string */
+PHP_FUNCTION(random_bytes)
+{
+	zend_long size;
+	zend_string *bytes;
+
+	ZEND_PARSE_PARAMETERS_START(1, 1)
+		Z_PARAM_LONG(size)
+	ZEND_PARSE_PARAMETERS_END();
+
+	if (size < 1) {
+		zend_argument_value_error(1, "must be greater than 0");
+		RETURN_THROWS();
+	}
+
+	bytes = zend_string_alloc(size, 0);
+
+	if (php_random_bytes_throw(ZSTR_VAL(bytes), size) == FAILURE) {
+		zend_string_release_ex(bytes, 0);
+		RETURN_THROWS();
+	}
+
+	ZSTR_VAL(bytes)[size] = '\0';
+
+	RETURN_STR(bytes);
+}
+/* }}} */
+
+/* {{{ Return an arbitrary pseudo-random integer */
+PHP_FUNCTION(random_int)
+{
+	zend_long min, max, result;
+
+	ZEND_PARSE_PARAMETERS_START(2, 2)
+		Z_PARAM_LONG(min)
+		Z_PARAM_LONG(max)
+	ZEND_PARSE_PARAMETERS_END();
+
+	if (min > max) {
+		zend_argument_value_error(1, "must be less than or equal to argument #2 ($max)");
+		RETURN_THROWS();
+	}
+
+	if (php_random_int_throw(min, max, &result) == FAILURE) {
+		RETURN_THROWS();
+	}
+
+	RETURN_LONG(result);
+}
+/* }}} */
+
+/* {{{ PHP_MINIT_FUNCTION */
+PHP_MINIT_FUNCTION(random)
+{
+	REGISTER_LONG_CONSTANT("MT_RAND_MT19937", MT_RAND_MT19937, CONST_CS | CONST_PERSISTENT);
+	REGISTER_LONG_CONSTANT("MT_RAND_PHP",     MT_RAND_PHP, CONST_CS | CONST_PERSISTENT);
+
+	RANDOM_G(random_fd) = -1;
+
+	return SUCCESS;
+}
+/* }}} */
+
+/* {{{ PHP_MSHUTDOWN_FUNCTION */
+PHP_MSHUTDOWN_FUNCTION(random)
+{
+	if (RANDOM_G(random_fd) > 0) {
+		close(RANDOM_G(random_fd));
+		RANDOM_G(random_fd) = -1;
+	}
+
+	return SUCCESS;
+}
+/* }}} */
+
+/* {{{ PHP_RINIT_FUNCTION */
+PHP_RINIT_FUNCTION(random)
+{
+#if defined(ZTS) && defined(COMPILE_DL_RANDOM)
+	ZEND_TSRMLS_CACHE_UPDATE();
+#endif
+
+	RANDOM_G(lcg_seeded) = 0;
+	RANDOM_G(mt_rand_is_seeded) = 0;
+	RANDOM_G(mt_rand_mode) = MT_RAND_MT19937;
+
+	return SUCCESS;
+}
+/* }}} */
+
+/* {{{ random_module_entry */
+zend_module_entry random_module_entry = {
+	STANDARD_MODULE_HEADER,
+	"random",					/* Extension name */
+	ext_functions,				/* zend_function_entry */
+	PHP_MINIT(random),			/* PHP_MINIT - Module initialization */
+	PHP_MSHUTDOWN(random),		/* PHP_MSHUTDOWN - Module shutdown */
+	PHP_RINIT(random),			/* PHP_RINIT - Request initialization */
+	NULL,						/* PHP_RSHUTDOWN - Request shutdown */
+	NULL,						/* PHP_MINFO - Module info */
+	PHP_VERSION,				/* Version */
+	STANDARD_MODULE_PROPERTIES
+};
+/* }}} */
+
+#ifdef COMPILE_DL_RANDOM
+# ifdef ZTS
+ZEND_TSRMLS_CACHE_DEFINE()
+# endif
+ZEND_GET_MODULE(random)
+#endif
