+++ ext/random/tests/lcg_value_basic.phpt
