+++ ext/random/tests/mt_getrandmax_basic.phpt
