+++ ext/random/tests/mt_srand_basic.phpt
