+++ ext/random/tests/random_bytes_error.phpt
