+++ ext/random/tests/srand_basic.phpt
