+++ ext/session/session.c
@@ -37,16 +37,15 @@
 #include "php_variables.h"
 #include "php_session.h"
 #include "session_arginfo.h"
-#include "ext/standard/php_random.h"
 #include "ext/standard/php_var.h"
 #include "ext/date/php_date.h"
-#include "ext/standard/php_lcg.h"
 #include "ext/standard/url_scanner_ex.h"
 #include "ext/standard/info.h"
 #include "zend_smart_str.h"
 #include "ext/standard/url.h"
 #include "ext/standard/basic_functions.h"
 #include "ext/standard/head.h"
+#include "ext/random/php_random.h"
 
 #include "mod_files.h"
 #include "mod_user.h"
