+++ ext/standard/basic_functions.h
@@ -27,6 +27,9 @@
 
 #include "url_scanner_ex.h"
 
+/* for MT_N */
+#include "ext/random/php_random.h"
+
 #if defined(_WIN32) && !defined(__clang__)
 #include <intrin.h>
 #endif
@@ -50,8 +53,6 @@ PHP_RSHUTDOWN_FUNCTION(browscap);
 PHPAPI int _php_error_log(int opt_err, const char *message, const char *opt, const char *headers);
 PHPAPI int _php_error_log_ex(int opt_err, const char *message, size_t message_len, const char *opt, const char *headers);
 PHPAPI int php_prefix_varname(zval *result, zend_string *prefix, const char *var_name, size_t var_name_len, bool add_underscore);
-
-#define MT_N (624)
 
 /* Deprecated type aliases -- use the standard types instead */
 typedef uint32_t php_uint32;
@@ -81,14 +82,6 @@ typedef struct _php_basic_globals {
 	/* filestat.c && main/streams/streams.c */
 	zend_string *CurrentStatFile, *CurrentLStatFile;
 	php_stream_statbuf ssb, lssb;
-
-	/* mt_rand.c */
-	uint32_t state[MT_N+1];  /* state vector + 1 extra to not violate ANSI C */
-	uint32_t *next;       /* next random value is computed from here */
-	int      left;        /* can *next++ this many times before reloading */
-
-	bool mt_rand_is_seeded; /* Whether mt_rand() has been seeded */
-	zend_long mt_rand_mode;
 
 	/* syslog.c */
 	char *syslog_device;
