+++ ext/standard/uniqid.c
@@ -31,8 +31,7 @@
 #include <sys/time.h>
 #endif
 
-#include "php_lcg.h"
-#include "php_random.h"
+#include "ext/random/php_random.h"
 
 #ifdef HAVE_GETTIMEOFDAY
 ZEND_TLS struct timeval prev_tv = { 0, 0 };
